Added LISS and EASY resources and split up the ingestion for every data provider.
