+++ .gitignore
@@ -168,6 +168,8 @@ cython_debug/
 
 # End of https://www.toptal.com/developers/gitignore/api/python
 /cbs-metadata/
+/easy-metadata/
+/liss-metadata/
 
 /resources/test-data
