+++ resources/templates/easy_dataverse_template.json
@@ -182,7 +182,6 @@
           }
         ]
       }
-    },
-    "license": "DANS Licence"
+    }
   }
 }
